requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.22
python-dotenv==1.0.1
gunicorn==21.2.0
pytest==7.4.0
